Refactor widgets into their own separate folders
- clean up the widgets folder into their respective folders to improve
readability and prevent stockpile of widgets in a single folder
