+++ lib/views/auth/forgot_password_screen.dart
@@ -4,10 +4,10 @@ import "package:flutter/material.dart";
 import 'package:go_router/go_router.dart';
 
 // Widgets
-import '../../widgets/custom_rich_text.dart';
-import '../../widgets/error_text.dart';
-import '../../widgets/login_signup_widgets.dart';
-import '../../widgets/logo.dart';
+import 'widgets/logo.dart';
+import 'widgets/login_signup_widgets.dart';
+import 'widgets/custom_rich_text.dart';
+import 'widgets/error_text.dart';
 
 // Utilities
 import '../../services/auth_services.dart';
