+++ lib/views/auth/widgets/custom_rich_text.dart
