+++ lib/views/auth/widgets/error_text.dart
