+++ lib/views/profile/widgets/profile_widget.dart
@@ -70,4 +70,3 @@ class CustomInkWell extends StatelessWidget {
     );
   }
 }
-
